+++ R/pairwise_comparisons.R
@@ -1,7 +1,8 @@
 #' @title Multiple pairwise comparison tests with tidy data
 #' @name pairwise_comparisons
-#' @description Calculate parametric, non-parametric, and robust pairwise
-#'   comparisons between group levels with corrections for multiple testing.
+#' @description Calculate parametric, non-parametric, robust, and Bayes Factor
+#'   pairwise comparisons between group levels with corrections for multiple
+#'   testing.
 #'
 #' @param data A dataframe from which variables specified are to be taken. A
 #'   matrix or tables will **not** be accepted.
@@ -38,6 +39,21 @@
 #'   columns across the different types of statistics, there will be additional
 #'   columns specific to the `type` of test being run.
 #'
+#'   This function provides a unified syntax to carry out pairwise comparison
+#'   tests and internally relies on other packages to carry out these tests. For
+#'   more details about the included tests, see the documentation for the
+#'   respective functions:
+#'   \itemize{
+#'   \item *parametric* : [stats::pairwise.t.test()] (paired) and
+#'   [PMCMRplus::gamesHowellTest()] (unpaired)
+#'   \item *non-parametric* :
+#'   [PMCMRplus::durbinAllPairsTest()] (paired) and
+#'   [PMCMRplus::kwAllPairsDunnTest()] (unpaired)
+#'   \item *robust* :
+#'   [WRS2::rmmcp()] (paired) and  [WRS2::lincon()] (unpaired)
+#'   \item *Bayes Factor* : [BayesFactor::ttestBF()]
+#'   }
+#'
 #'   The `significance` column asterisks indicate significance levels of
 #'   *p*-values in the American Psychological Association (APA) mandated format:
 #'   \itemize{
@@ -47,20 +63,15 @@
 #'   \item `***` : < 0.001
 #'   }
 #'
-#' @importFrom dplyr select rename mutate everything full_join vars mutate_if
-#' @importFrom dplyr bind_cols rename_all recode matches rowwise ungroup
-#' @importFrom stats p.adjust pairwise.t.test na.omit aov TukeyHSD
+#' @importFrom dplyr select rename mutate everything mutate_if across starts_with
+#' @importFrom dplyr bind_cols matches rowwise ungroup
+#' @importFrom stats p.adjust pairwise.t.test na.omit aov
 #' @importFrom WRS2 lincon rmmcp
-#' @importFrom tidyr gather spread separate
-#' @importFrom dunn.test dunn.test
-#' @importFrom PMCMRplus durbinAllPairsTest
+#' @importFrom PMCMRplus durbinAllPairsTest kwAllPairsDunnTest gamesHowellTest
 #' @importFrom rlang !! enquo as_string ensym
-#' @importFrom forcats fct_relabel
 #' @importFrom purrr map2 map_dfr
-#' @importFrom broomExtra tidy easystats_to_tidy_names
-#' @importFrom ipmisc stats_type_switch
-#' @importFrom tidyBF bf_ttest
-#' @importFrom utils capture.output
+#' @importFrom ipmisc stats_type_switch specify_decimal_p long_to_wide_converter
+#' @importFrom ipmisc signif_column
 #'
 #' @examples
 #'
@@ -185,8 +196,7 @@ pairwise_comparisons <- function(data,
   type <- ipmisc::stats_type_switch(type)
 
   # ensure the arguments work quoted or unquoted
-  x <- rlang::ensym(x)
-  y <- rlang::ensym(y)
+  c(x, y) %<-% c(rlang::ensym(x), rlang::ensym(y))
 
   # ---------------------------- data cleanup -------------------------------
 
@@ -204,164 +214,25 @@ pairwise_comparisons <- function(data,
   x_vec <- df_internal %>% dplyr::pull({{ x }})
   y_vec <- df_internal %>% dplyr::pull({{ y }})
 
-  # ---------------------------- parametric ---------------------------------
-
-  if (type %in% c("parametric", "bayes")) {
-    if (isTRUE(var.equal) || isTRUE(paired)) {
-      # anova model
-      aovmodel <- stats::aov(
-        formula = rlang::new_formula({{ y }}, {{ x }}),
-        data = df_internal
-      )
-
-      # safeguarding against edge cases
-      aovmodel$model %<>%
-        dplyr::mutate(
-          .data = .,
-          {{ x }} := forcats::fct_relabel(
-            .f = {{ x }},
-            .fun = ~ gsub(x = .x, pattern = "-", replacement = "_")
-          )
-        )
-
-      # extracting and cleaning up Tukey's HSD output
-      df_tukey <-
-        stats::TukeyHSD(x = aovmodel, conf.level = 0.95) %>%
-        broomExtra::tidy(.) %>%
-        dplyr::select(.data = ., comparison = contrast, mean.difference = estimate) %>%
-        tidyr::separate(
-          data = .,
-          col = comparison,
-          into = c("group1", "group2"),
-          sep = "-"
-        ) %>%
-        dplyr::mutate_at(
-          .tbl = .,
-          .vars = dplyr::vars(dplyr::matches("^group[0-9]$")),
-          .funs = ~ gsub(x = ., pattern = "_", replacement = "-")
-        )
-
-      # tidy dataframe with results from pairwise tests
-      df_tidy <-
-        stats::pairwise.t.test(
-          x = y_vec,
-          g = x_vec,
-          p.adjust.method = "none",
-          paired = paired,
-          alternative = "two.sided",
-          na.action = na.omit
-        ) %>%
-        broomExtra::tidy(.) %>%
-        p_adjust_column_adder(df = ., p.adjust.method = p.adjust.method)
-
-      # combining mean difference and results from pairwise t-test
-      df <-
-        dplyr::full_join(x = df_tukey, y = df_tidy, by = c("group1", "group2")) %>%
-        dplyr::rename(.data = ., group2 = group1, group1 = group2)
-
-      # test details
-      test.details <- "Student's t-test"
-    } else {
-      # dataframe with Games-Howell test results
-      df <-
-        games_howell(data = df_internal, x = {{ x }}, y = {{ y }}) %>%
-        p_adjust_column_adder(df = ., p.adjust.method = p.adjust.method) %>%
-        dplyr::select(.data = ., -conf.low, -conf.high)
-
-      # test details
-      test.details <- "Games-Howell test"
-    }
-  }
-
-  # ---------------------------- bayes factor --------------------------------
-
-  # print a message telling the user that this is currently not supported
-  if (type == "bayes") {
-    # creating a list of dataframes with subsections of data
-    df_list <-
-      purrr::map2(
-        .x = as.character(df$group1),
-        .y = as.character(df$group2),
-        .f = function(a, b) droplevels(dplyr::filter(df_internal, {{ x }} %in% c(a, b)))
-      )
-
-    # combining results into a single dataframe and returning it
-    df_tidy <-
-      purrr::map_dfr(
-        .x = df_list,
-        .f = ~ tidyBF::bf_ttest(
-          data = .x,
-          x = {{ x }},
-          y = {{ y }},
-          paired = paired,
-          bf.prior = bf.prior,
-          output = "results"
-        )
-      ) %>%
-      dplyr::rowwise() %>%
-      dplyr::mutate(label = paste0(
-        "list(~log[e](BF[10])", "==", specify_decimal_p(x = log_e_bf10, k = k), ")"
-      )) %>%
-      dplyr::ungroup() %>%
-      dplyr::mutate(.data = ., test.details = "Student's t-test")
-
-    # early return (no further cleanup required)
-    return(
-      dplyr::mutate_if(
-        .tbl = dplyr::bind_cols(dplyr::select(df, group1, group2), df_tidy),
-        .predicate = is.factor,
-        .funs = ~ as.character(.)
-      )
-    )
-  }
-
   # ---------------------------- nonparametric ----------------------------
 
   if (type == "nonparametric") {
+    # # running Dunn test
     if (isFALSE(paired)) {
-      # running Dunn test
-      invisible(utils::capture.output(df <-
-        as.data.frame(
-          dunn.test::dunn.test(
-            x = y_vec,
-            g = x_vec,
-            table = FALSE,
-            kw = FALSE,
-            label = FALSE,
-            alpha = 0.05,
-            method = "none",
-            altp = TRUE
-          )
-        ), file = NULL))
-
-      # cleanup
-      df %<>%
-        as_tibble(.) %>%
-        broomExtra::easystats_to_tidy_names(.) %>%
-        dplyr::mutate(z.value = abs(statistic)) %>%
-        dplyr::select(
-          .data = .,
-          comparisons,
-          z.value,
-          "p.value" = "altp.adjusted",
-          dplyr::everything(),
-          -c(chi2:altp)
-        ) %>%
-        tidyr::separate(
-          data = .,
-          col = "comparisons",
-          into = c("group1", "group2"),
-          sep = " - ",
-          remove = TRUE
-        )
+      mod <-
+        suppressWarnings(PMCMRplus::kwAllPairsDunnTest(
+          x = y_vec,
+          g = x_vec,
+          na.action = na.omit,
+          p.adjust.method = "none"
+        ))
 
       # test details
       test.details <- "Dunn test"
     }
 
-    # converting the entered long format data to wide format
+    # # running Durbin-Conover test
     if (isTRUE(paired)) {
-      # creating model object
       mod <-
         PMCMRplus::durbinAllPairsTest(
           y = na.omit(matrix(
@@ -372,42 +243,27 @@ pairwise_comparisons <- function(data,
           p.adjust.method = "none"
         )
 
-      # combining into one dataframe
-      df <-
-        dplyr::bind_cols(
-          matrix_to_tidy(m = mod$statistic, col_names = c("group2", "group1", "W")),
-          dplyr::select(
-            matrix_to_tidy(m = mod$p.value, col_names = c("group2", "group1", "p.value")),
-            -dplyr::contains("group")
-          )
-        )
-
       # test details
       test.details <- "Durbin-Conover test"
     }
 
     # cleanup
-    df %<>% p_adjust_column_adder(df = ., p.adjust.method = p.adjust.method)
+    df <- PMCMR_to_tibble(mod)
   }
 
   # ---------------------------- robust ----------------------------------
 
   if (type == "robust") {
+    # extracting the robust pairwise comparisons
     if (isFALSE(paired)) {
-      # object with all details about pairwise comparisons
-      rob_pairwise_df <-
+      mod <-
         WRS2::lincon(
           formula = rlang::new_formula({{ y }}, {{ x }}),
           data = df_internal,
           tr = tr
         )
-    }
-
-    # converting to long format and then getting it back in wide so that the
-    # rowid variable can be used as the block variable
-    if (isTRUE(paired)) {
-      # running pairwise multiple comparison tests
-      rob_pairwise_df <-
+    } else {
+      mod <-
         WRS2::rmmcp(
           y = df_internal[[rlang::as_name(y)]],
           groups = df_internal[[rlang::as_name(x)]],
@@ -416,73 +272,120 @@ pairwise_comparisons <- function(data,
         )
     }
 
-    # extracting the robust pairwise comparisons and tidying up names
-    rob_df_tidy <-
-      suppressMessages(as_tibble(rob_pairwise_df$comp, .name_repair = "unique")) %>%
-      dplyr::rename(group1 = Group...1, group2 = Group...2)
-
     # cleaning the raw object and getting it in the right format
     df <-
-      dplyr::full_join(
-        # dataframe comparing comparison details
-        x = p_adjust_column_adder(df = rob_df_tidy, p.adjust.method = p.adjust.method) %>%
-          tidyr::gather(
-            data = .,
-            key = "key",
-            value = "rowid",
-            group1:group2
-          ),
-        # dataframe with factor levels
-        y = enframe(x = rob_pairwise_df$fnames, name = "rowid"),
-        by = "rowid"
-      ) %>%
-      dplyr::select(.data = ., -rowid) %>%
-      tidyr::spread(data = ., key = "key", value = "value")
+      suppressMessages(as_tibble(mod$comp, .name_repair = "unique")) %>%
+      dplyr::rename(group1 = Group...1, group2 = Group...2) %>%
+      dplyr::mutate(dplyr::across(
+        .cols = dplyr::starts_with("group"),
+        .fns = ~ as.character(setNames(mod$fnames, seq_along(mod$fnames))[as.character(.)])
+      ))
 
     # for paired designs, there will be an unnecessary column to remove
     if (("p.crit") %in% names(df)) df %<>% dplyr::select(.data = ., -p.crit)
 
     # renaming confidence interval names
-    df %<>% dplyr::rename(.data = ., conf.low = ci.lower, conf.high = ci.upper)
+    df %<>% dplyr::rename(estimate = psihat, conf.low = ci.lower, conf.high = ci.upper)
 
     # test details
     test.details <- "Yuen's trimmed means test"
   }
 
-  # ---------------------------- cleanup ----------------------------------
+  # ---------------------------- parametric ---------------------------------
 
-  # formatting label
-  adjust_text <- ifelse(p.adjust.method == "none", "unadjusted", "adjusted")
+  if (type %in% c("parametric", "bayes")) {
+    if (isTRUE(var.equal) || isTRUE(paired)) {
+      # tidy dataframe with results from pairwise tests
+      df <-
+        stats::pairwise.t.test(
+          x = y_vec,
+          g = x_vec,
+          p.adjust.method = "none",
+          paired = paired,
+          na.action = na.omit
+        ) %>%
+        .$p.value %>%
+        matrix_to_tidy(., "p.value")
+
+      # test details
+      test.details <- "Student's t-test"
+    } else {
+      # anova model
+      aovmodel <- stats::aov(rlang::new_formula({{ y }}, {{ x }}), df_internal)
+
+      # dataframe with Games-Howell test results
+      df <- PMCMR_to_tibble(PMCMRplus::gamesHowellTest(aovmodel, p.adjust.method = "none"))
+
+      # test details
+      test.details <- "Games-Howell test"
+    }
+  }
+
+  # ---------------------------- bayes factor --------------------------------
+
+  if (type == "bayes") {
+    # combining results into a single dataframe and returning it
+    df_tidy <-
+      purrr::map_dfr(
+        # creating a list of dataframes with subsections of data
+        .x = purrr::map2(
+          .x = as.character(df$group1),
+          .y = as.character(df$group2),
+          .f = function(a, b) droplevels(dplyr::filter(df_internal, {{ x }} %in% c(a, b)))
+        ),
+        # internal function to carry out BF t-test
+        .f = ~ bf_internal_ttest(
+          data = .x,
+          x = {{ x }},
+          y = {{ y }},
+          paired = paired,
+          bf.prior = bf.prior
+        )
+      ) %>%
+      dplyr::rowwise() %>%
+      dplyr::mutate(label = paste0(
+        "list(~log[e](BF[10])", "==", specify_decimal_p(x = log_e_bf10, k = k), ")"
+      )) %>%
+      dplyr::ungroup() %>%
+      dplyr::mutate(.data = ., test.details = "Student's t-test")
+
+    # combine it with the other details
+    df <- dplyr::bind_cols(dplyr::select(df, group1, group2), df_tidy)
+  }
+
+  # ---------------------------- cleanup ----------------------------------
 
   # final cleanup for p-value labels
   df %<>%
-    dplyr::mutate_if(
-      .tbl = .,
-      .predicate = is.factor,
-      .funs = ~ as.character(.)
-    ) %>%
-    dplyr::rowwise() %>%
-    dplyr::mutate(label = specify_decimal_p(x = p.value, k = k, p.value = TRUE)) %>%
-    dplyr::mutate(
-      label = dplyr::case_when(
-        label == "< 0.001" ~ paste0("list(~italic(p)[", adjust_text, "]<=", "0.001", ")"),
-        TRUE ~ paste0("list(~italic(p)[", adjust_text, "]==", label, ")")
-      )
-    ) %>%
-    dplyr::mutate(
-      test.details = test.details,
-      p.value.adjustment = p_adjust_text(p.adjust.method)
-    ) %>%
-    dplyr::select(.data = ., group1, group2, dplyr::everything()) %>%
-    dplyr::arrange(group1, group2)
+    dplyr::mutate_if(.tbl = ., .predicate = is.factor, .funs = ~ as.character(.)) %>%
+    dplyr::arrange(group1, group2) %>%
+    dplyr::select(.data = ., group1, group2, dplyr::everything())
+
+  # clean-up for non-Bayes tests
+  if (type != "bayes") {
+    df %<>%
+      dplyr::mutate(p.value = stats::p.adjust(p = p.value, method = p.adjust.method)) %>%
+      signif_column(data = ., p = p.value) %>%
+      dplyr::mutate(
+        test.details = test.details,
+        p.value.adjustment = p_adjust_text(p.adjust.method)
+      ) %>%
+      dplyr::rowwise() %>%
+      dplyr::mutate(
+        label = dplyr::case_when(
+          p.value.adjustment != "None" ~ paste0(
+            "list(~italic(p)[",
+            p.value.adjustment,
+            "-corrected]==",
+            specify_decimal_p(p.value, k, TRUE),
+            ")"
+          ),
+          TRUE ~ paste0("list(~italic(p)[uncorrected]==", specify_decimal_p(p.value, k, TRUE), ")")
+        )
+      ) %>%
+      dplyr::ungroup()
+  }
 
   # return
-  return(dplyr::ungroup(df))
+  return(df)
 }
-
-
-#' @name pairwise_comparisons
-#' @aliases  pairwise_comparisons
-#' @export
-
-pairwise_p <- pairwise_comparisons